requirements.txt
google-genai
pillow
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyPDF2
edge-tts
pytrends
huggingface_hub
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9